Fix mutable default args
I guessed at types for the optional args and a few variables that mypy needed type annotations for.

As we fill in types, keep in mind that `typing` also has duck types like `Iterable` and `Sequence`. There's also `Mapping` although if the code expects every map to be a `dict` then we might as well stick with `dict` or `Dict[k, v]` types.