Make Engine Handle Flows with Nested Dictionaries
Previously, Engine expected a flow to be a flat dictionary with process
names as keys. This requires that processes have unique names, which is
not a general requirement of Vivarium (e.g. two cells can have processes
of the same name). This commit fixes this behavior to support nested
flows.