Use initial instead of mother state for division
Division used the mother cell state as the default state and deep-merged
the result of running the dividers into that default. This led to
daughter cells getting molecules that were sent by the dividers to the
other daughter.

Instead, we should use the result of running the dividers as the default
daughter state. Then, we can deep-merge in any state provided by the
meta-division process as part of the `_divide` update. This commit makes
these changes.