Store ParallelProcess Objs Alongside Process Objs
Instead of holding ParallelProcess objects in a separate Engine.parallel
dictionary, put them in place of their associated Process objects in
Engine.processes, Engine.process_paths, and Engine.state. This also
involves removing InvokeProcess and the `invoke` keyword from the Engine
constructor. While these are public, the `invoke` keyword was not
documented, and use of InvokeProcess is not discussed in our
documentation. Therefore, neither are part of our supported API, so
removing them does not constitute a breaking change.