Parallelize Processes After Division
With the move to Process Commands, we stopped parallelizing processes in
`Engine._invoke_process` and instead performed parallelization in the
constructor. This had the unintended effect of not parallelizing
processes with the `parallel` attribute if they were produced by
division. This commit ensures that these processes are parallelized as
well.