Support equality checks in deep_merge_check
Before, `deep_merge_check()` only supported checking for conflicts using
`is`. This made sense when Numpy arrays or complex objects might be
involved, but for parallel emits, we want to do equality checks so that,
for example, `[1]` and `[1]` aren't considered to be conflicting values.
This commit adds a flag to let callers choose to do equality checks with
`==` instead. It also makes the error messages easier to understand.