Convert units to match Store.units after update
When a `Store` has specified units, we want `Store.value` to always be
in those units. This ensures that inner and outer simulations use the
same units when emitting data because when we pass Store configs between
simulations, those configs contain `_value` with associated units. This
also makes inspecting the simulation state nicer.